PR3 commit 4: duels gear wear (Q-0054), PIL cards, live fixes
- Q-0054 CLOSED: a finished PvP deathmatch ticks each human fighter's
  weapon + armor once (ACTION_DUEL wear plan; wear notes surface on the
  final duel embed; timeout path included; bot fighters skipped — they
  grant nothing, so no farming vector and nothing to wear).
- Visuals (owner decision: inventory card + stat card): the existing
  tested inventory-card renderer finally gets its first caller (hub
  Inventory button sends an ephemeral PNG, embed always remains); NEW
  StatCardSpec/render_stat_card in utils/mining_render.py — !character
  attaches the §7.6 zero-art stat card (level/XP bar + location +
  deepest + gear condition + wealth), embed fallback when PIL is
  unavailable. Stale 'Pillow not in requirements' docstring fixed
  (it has been pinned since the BTD6 map work).
- Live-DB catch: add_game_xp's $4 fed both a BIGINT and an INTEGER
  column — asyncpg AmbiguousParameterError invisible to mock tests;
  explicit ::bigint/::int casts, SQL pins updated.

Live-verified end-to-end on real Postgres: mine→3 XP, craft→8 XP,
forced level-up renders the inline note, diamond lantern descends to
the Magma core (depth 3/3 — previously unreachable), depth records
award once and never re-award, magma mining earns depth-scaled XP (6),
top_total_xp leaderboard reads correctly.

+++ disbot/utils/db/games/game_xp.py
@@ -54,13 +54,15 @@ async def add_game_xp(
     One upsert with day-rollover: ``day_xp`` accumulates while ``day``
     matches and restarts at *amount* on a new day.
     """
+    # $4 feeds both xp (BIGINT) and day_xp (INTEGER) — the explicit casts
+    # keep asyncpg's parameter-type deduction unambiguous.
     row = await pool.fetchone(
         """INSERT INTO game_xp (user_id, guild_id, game, xp, day, day_xp)
-           VALUES ($1, $2, $3, $4, $5, $4)
+           VALUES ($1, $2, $3, $4::bigint, $5, $4::int)
            ON CONFLICT (user_id, guild_id, game) DO UPDATE SET
-               xp = game_xp.xp + $4,
+               xp = game_xp.xp + $4::bigint,
                day_xp = CASE WHEN game_xp.day = $5
-                             THEN game_xp.day_xp + $4 ELSE $4 END,
+                             THEN game_xp.day_xp + $4::int ELSE $4::int END,
                day = $5,
                updated_at = now()
            RETURNING xp""",
